# [quantization] Use more efficient specialized Quantize routine