--- a/aten/src/ATen/cpu/vec256/vec256_qint.h
+++ b/aten/src/ATen/cpu/vec256/vec256_qint.h
@@ -40,7 +40,24 @@ namespace {
 #if defined(__AVX__) && !defined(_MSC_VER)
 
 template <typename T>
-void QuantizeAvx2(
+__m256i pack_saturate_and_clamp(__m256i first, __m256i second, T min_val, T max_val);
+
+template <>
+__m256i pack_saturate_and_clamp<int8_t>(__m256i first, __m256i second, int8_t min_val, int8_t max_val) {
+    __m256i packed_and_sat = _mm256_packs_epi16(first, second);
+    return _mm256_max_epi8(_mm256_set1_epi8(min_val),
+                _mm256_min_epi8(packed_and_sat, _mm256_set1_epi8(max_val)));
+}
+
+template <>
+__m256i pack_saturate_and_clamp<uint8_t>(__m256i first, __m256i second, uint8_t min_val, uint8_t max_val) {
+    __m256i packed_and_sat = _mm256_packus_epi16(first, second);
+    return _mm256_max_epu8(_mm256_set1_epi8(min_val),
+                _mm256_min_epu8(packed_and_sat, _mm256_set1_epi8(max_val)));
+}
+
+template <typename T>
+inline void __attribute__((always_inline)) QuantizeAvx2(
     const float* src,
     typename T::underlying* dst,
     int len,
@@ -48,64 +65,48 @@ void QuantizeAvx2(
     int64_t zero_point) {
 #if defined(__AVX2__) && defined(__FMA__)
   constexpr int VLEN = 8;
-  constexpr float min_val = std::numeric_limits<typename T::underlying>::min();
-  constexpr float max_val = std::numeric_limits<typename T::underlying>::max();
+  constexpr auto min_val = std::numeric_limits<typename T::underlying>::min();
+  constexpr auto max_val = std::numeric_limits<typename T::underlying>::max();
   std::size_t i = 0;
   __m256 inverse_scale_v = _mm256_set1_ps(1.f / scale);
-  __m256i shuffle_mask_v = _mm256_set_epi8(
-      0xff,
-      0xff,
-      0xff,
-      0xff,
-      0xff,
-      0xff,
-      0xff,
-      0xff,
-      0xff,
-      0xff,
-      0xff,
-      0xff,
-      0x0c,
-      0x08,
-      0x04,
-      0x00,
-      0xff,
-      0xff,
-      0xff,
-      0xff,
-      0xff,
-      0xff,
-      0xff,
-      0xff,
-      0xff,
-      0xff,
-      0xff,
-      0xff,
-      0x0c,
-      0x08,
-      0x04,
-      0x00);
   __m256i permute_mask_v =
-      _mm256_set_epi32(0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x04, 0x00);
-  for (; i < len / VLEN * VLEN; i += VLEN) {
-    __m256 src_v = _mm256_loadu_ps(src + i);
-    __m256 transformed_v = _mm256_fmadd_ps(
-        src_v, inverse_scale_v, _mm256_set1_ps(zero_point));
-    __m256 clipped_v = _mm256_min_ps(
-        _mm256_max_ps(transformed_v, _mm256_set1_ps(min_val)),
-        _mm256_set1_ps(max_val));
-    __m256i rounded_v = _mm256_cvtps_epi32(clipped_v);
+      _mm256_set_epi32(0x07, 0x03, 0x06, 0x02, 0x05, 0x01, 0x04, 0x00);
+  int len_aligned = len / (VLEN * 4) * (VLEN * 4);
+  for (; i < len_aligned; i += 4 * VLEN) {
+    // x
+    __m256 x_vals = _mm256_load_ps(src + i);
+    __m256 x_transformed_v = _mm256_fmadd_ps(
+        x_vals, inverse_scale_v, _mm256_set1_ps(zero_point));
+    // y
+    __m256 y_vals = _mm256_load_ps(src + i + VLEN);
+    __m256 y_transformed_v = _mm256_fmadd_ps(
+        y_vals, inverse_scale_v, _mm256_set1_ps(zero_point));
+    // z
+    __m256 z_vals = _mm256_load_ps(src + i + 2 * VLEN);
+    __m256 z_transformed_v = _mm256_fmadd_ps(
+        z_vals, inverse_scale_v, _mm256_set1_ps(zero_point));
+    // w
+    __m256 w_vals = _mm256_load_ps(src + i + 3 * VLEN);
+    __m256 w_transformed_v = _mm256_fmadd_ps(
+        w_vals, inverse_scale_v, _mm256_set1_ps(zero_point));
 
-    // An instruction sequence to save 8 32-bit integers as 8 8-bit integers
-    rounded_v = _mm256_shuffle_epi8(rounded_v, shuffle_mask_v);
-    rounded_v = _mm256_permutevar8x32_epi32(rounded_v, permute_mask_v);
-    _mm_storel_epi64(
-        reinterpret_cast<__m128i*>(dst + i), _mm256_castsi256_si128(rounded_v));
+    __m256i x_rounded_v = _mm256_cvtps_epi32(x_transformed_v);
+    __m256i y_rounded_v = _mm256_cvtps_epi32(y_transformed_v);
+    __m256i z_rounded_v = _mm256_cvtps_epi32(z_transformed_v);
+    __m256i w_rounded_v = _mm256_cvtps_epi32(w_transformed_v);
+
+    __m256i xy_packed_v = _mm256_packs_epi32(x_rounded_v, y_rounded_v);
+    __m256i zw_packed_v = _mm256_packs_epi32(z_rounded_v, w_rounded_v);
+    __m256i xyzw_clamped_v = pack_saturate_and_clamp<typename T::underlying>(xy_packed_v, zw_packed_v, min_val, max_val);
+
+    xyzw_clamped_v =
+        _mm256_permutevar8x32_epi32(xyzw_clamped_v, permute_mask_v);
+    _mm256_storeu_si256(reinterpret_cast<__m256i*>(dst + i), xyzw_clamped_v);
   }
 
   for (; i < len; ++i) {
     float transformed = zero_point + src[i] / scale;
-    float clipped = std::min(std::max(transformed, min_val), max_val);
+    float clipped = std::min(std::max(transformed, float(min_val)), float(max_val));
     // Not exactly the same behavior as the vectorized code.
     // The vectorized code above always rounds to even in halfway cases
     // (https://software.intel.com/en-us/node/523819), but std::nearbyint